Adding M16C support
git-svn-id: svn://svn.code.sf.net/p/nuttx/code/trunk@1486 42af7a65-404d-4744-a932-0658087f49c3

+++ arch/sh/include/m16c/types.h
@@ -72,7 +72,7 @@ typedef unsigned long long uint64;
  * irqsave()
  */
 
-typedef unsigned int irqstate_t;
+typedef uint16 irqstate_t;
 
 #endif /* __ASSEMBLY__ */
